Optimized the memory and wall clock cost of the processing stages.
-- Fused the Demons force computation into a single Numba kernel whose arithmetic order matches the operator chain it replaces, removing 22 full-field temporaries per image pair while holding the result bit-identical.
-- Accumulated the groupwise deformations through raw field buffers, so every contribution after the first writes through an existing buffer instead of allocating a fresh pair of fields.
-- Tabulated the B-spline column coefficients once per field rather than once per pixel, and allocated the sampled deformation fields without a zero fill the sampler overwrites in full.
-- Reused one float32 destination across every extraction batch, which the binary's int16 frames convert into once per chunk.
-- Removed the rigid shift the two-photon path applied twice, by sharing one buffer between the registered frames and the smoothed copy the nonrigid correlation reads.
-- Cached the soma-masked coordinates and the soma pixel count that three separate ROI statistics each derived from the mask independently.
-- Resolved each neuropil mask over its expanded region rather than over a frame-sized boolean image, and counted shared pixels through a merge over lists sorted once per ROI.
-- Guarded every array loader cast, so a file already holding the declared width loads without a copy, and dropped the casts whose own operands pin the width.
-- Held the reported array summary to a chunked two-pass accumulator, which lowers its peak from the whole file to one block and raises its accumulation width to float64.
-- Narrowed the multi-recording memory estimator to the geometry it reads, and carried each recording root beside its geometry so the estimate resolves the recording its specifier names.
-- Replaced the quadratic plane specifier scan in the ready-job test with a parse of the specifier.
-- Allocated the viewer background stack at its final width, and reduced the per-ROI standard deviation to a single pass over the binned traces.
-- Squared the frame differences in place, folded the neuropil subtraction into one buffer, and clipped the registration write-back through the batch buffer it already holds.
-- Added regression tests pinning the fused kernel against its operator chain, the denoise accumulation order across worker counts, the shared-pixel count against the set intersection, and the estimator fallbacks.

+++ src/cindra/gui/overlays.py
@@ -78,7 +78,7 @@ def build_views(
     Returns:
         Array of shape (6, frame_height, frame_width, 3) containing RGB background views.
     """
-    views = np.zeros((len(BackgroundView), frame_height, frame_width, 3), dtype=np.float32)
+    views = np.zeros((len(BackgroundView), frame_height, frame_width, 3), dtype=np.uint8)
 
     for view_index in range(len(BackgroundView)):
         image = _build_single_view(
@@ -101,7 +101,7 @@ def build_views(
         image_uint8 = (image * 255).astype(np.uint8)
         views[view_index] = np.tile(image_uint8[:, :, np.newaxis], reps=(1, 1, 3))
 
-    return views.astype(np.uint8)
+    return views
 
 
 def display_views(
